Add ability to refresh main entries list; make the rest of MainActivity's vars private
Add necessary methods and callbacks to implement a crude refresh of
the entries list view upon creation of a new entry. Basically, drop
the adapter holding the current views and make a new one. Probably
not the best way to do it, but it works. Also, bump version to 0.2,
update changelog and to do lists.

Make the rest of MainActivity's vars private and add getters. Setters
not needed, since the vars are static final and do not need to be
changed.